Remove ttsim_params and include in ttsim_objects.
